Use attribute instead of assuming always 20
I do this mostly so I can properly test the healing. I was testing with some larger xp orbs (so that the difference is more obvious) but of course to see the difference i need to have the HP to handle it.

As a side effect, Zero can make an ARH boss that wears DAura armor and gains xp whenever a player within 100 blocks gains xp.