Split splinter validation for location and type.
I only need to check the type at the very beginning, since the Runnable would always be using the same type.

I didn't do it this way before because I didn't realize "don't change target-block-type after splintering has begun" was something I should consider.

I switched to is-invalid instead of is-valid since every usage was using the not-form. (My IDE was complaining lol)

(This commit is building up to a bigger refactor.)